Progress moving tests to behaviors
